+++ src/events/cps-after-close.ts
@@ -1,0 +1,9 @@
+type CpsAfterCloseEvent = CustomEvent<Record<PropertyKey, never>>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'cps-after-close': CpsAfterCloseEvent;
+  }
+}
+
+export default CpsAfterCloseEvent;