+++ src/events/cps-initial-focus.ts
@@ -1,0 +1,9 @@
+type CpsInitialFocus = CustomEvent<Record<PropertyKey, never>>;
+
+declare global {
+  interface GlobalEventHandlersEventMap {
+    'cps-initial-focus': CpsInitialFocus;
+  }
+}
+
+export default CpsInitialFocus;