+++ src/events/index.ts
@@ -1,3 +1,4 @@
+export { default as CpsAfterClose } from './cps-after-close.js';
 export { default as CpsAfterHide } from './cps-after-hide.js';
 export { default as CpsAfterShow } from './cps-after-show.js';
 export { default as CpsBlurEvent } from './cps-blur.js';
@@ -7,11 +8,13 @@ export { default as CpsCloseEvent } from './cps-close.js';
 export { default as CpsErrorEvent } from './cps-error.js';
 export { default as CpsFocusEvent } from './cps-focus.js';
 export { default as CpsHide } from './cps-hide.js';
+export { default as CpsInitialFocus } from './cps-initial-focus.js';
 export { default as CpsInputEvent } from './cps-input.js';
 export { default as CpsInvalidEvent } from './cps-invalid.js';
 export { default as CpsLoadEvent } from './cps-load.js';
 export { default as CpsRemoveEvent } from './cps-remove.js';
 export { default as CpsReposition } from './cps-reposition.js';
+export { default as CpsRequestClose } from './cps-request-close.js';
 export { default as CpsSelect } from './cps-select.js';
 export { default as CpsShow } from './cps-show.js';
 export { default as CpsTabHide } from './cps-tab-hide.js';